# src/main/java/com/example/api/contracts/controller/ContractReviewController.java
package com.example.api.contracts.controller;

import com.example.api.announcement.dto.PageNumberRequest;
import com.example.api.contracts.ReviewQueryService;
import com.example.api.contracts.ContractReviewService;
import com.example.api.contracts.dto.AddReviewCommand;
import com.example.api.contracts.dto.DeleteReviewRequest;
import com.example.api.contracts.dto.QueryEmployersReviewCommand;
import com.example.api.review.dto.ModifyReviewRequest;
import com.example.api.review.dto.ReviewAvailableResponse;
import com.example.api.review.dto.ReviewResponse;
import jakarta.validation.Valid;
import jakarta.validation.constraints.NotNull;
import java.util.List;
import lombok.RequiredArgsConstructor;
import org.hibernate.validator.constraints.Range;
import org.springframework.http.ResponseEntity;
import org.springframework.security.core.annotation.AuthenticationPrincipal;
import org.springframework.web.bind.annotation.*;

@RestController
@RequiredArgsConstructor
@RequestMapping("/api/v1/contracts")
public class ContractReviewController {
    private final ContractReviewService contractReviewService;
    private final ReviewQueryService reviewQueryService;

    /**
     * @param requestMemberId 고용자 ( employer ID ) 가 본인이 작성한 리뷰 목록 가져오기
     * @return
     */
    @GetMapping("/review/my/employer")
    public ResponseEntity<List<ReviewResponse>> getMyReview(
            @AuthenticationPrincipal final Long requestMemberId,
            @RequestParam final Integer page
    ) {
        final QueryEmployersReviewCommand command = new QueryEmployersReviewCommand(requestMemberId);
        return ResponseEntity.ok(reviewQueryService.loadReviewsWithEmployerId(command, new PageNumberRequest(page)));
    }

    @PostMapping("/review")
    public ResponseEntity<String> addReview(
            @RequestBody @Valid final AddReviewRequest request,
            @AuthenticationPrincipal final Long memberId
    ) {
        final AddReviewCommand command = request.toCommand(memberId);
        contractReviewService.saveReview(command);
        return ResponseEntity.ok("리뷰가 성공적으로 작성되었습니다.");
    }

    record AddReviewRequest(
            @NotNull
            Long contractId,
            @NotNull
            Long businessId,
            @NotNull
            Long employeeId,
            @Range(min = 0, max = 5)
            Integer reviewScore,
            String reviewContent
    ) {
        AddReviewCommand toCommand(final Long requestMemberId) {
            return new AddReviewCommand(requestMemberId, businessId, employeeId, contractId, reviewContent, reviewScore);
        }
    }

    @PutMapping("/review/modify") // 리뷰 수정
    public ResponseEntity<ReviewResponse> modifyReview(
            @RequestBody ModifyReviewRequest modifyReviewRequest
    ) {
         ReviewResponse updatedReview = contractReviewService.modifyReview(modifyReviewRequest);
        return ResponseEntity.ok(updatedReview);
    }

    @DeleteMapping("/review/delete") // 리뷰 삭제
    public ResponseEntity<String> deleteReview(
            @RequestParam final Long reviewId
    ) {
        contractReviewService.deleteReview(new DeleteReviewRequest(reviewId));
        return ResponseEntity.ok("리뷰 삭제하였습니다");
    }
}