# src/main/resources/static/swagger.yaml
openapi: 3.0.1
info:
  title: Spring API
  description: 단팥 백엔드 API 명세서입니다.
  version: 1.0.0
servers:
  - url: https://www.danpat.store
#  - url: http://localhost:8080
security:
  - JWT_TOKEN: []
paths:
...
  /api/v1/contracts/review:
    post:
      tags:
        - ContractReviewController
      summary: "리뷰 작성하기"
      description: "리뷰 작성하기"
      operationId: addReview
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddReviewRequest"
            example:
              contractId: 9
              businessId: 9
              employeeId: 9
              reviewScore: 5
              reviewContent: "일 잘하네요"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
              example: "리뷰가 성공적으로 작성되었습니다."
  /api/v1/contracts/review/my/employer:
    get:
      tags:
        - ContractReviewController
      summary: "사장 본인이 쓴 리뷰 보기"
      description: "사장 본인이 쓴 리뷰 보기"
      operationId: getMyReview
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int64
            example: 1
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReviewResponse"
              example:
                - reviewId: 1
                  businessName: "테스트 사업장"
                  businessId: 10
                  employeeId: 3
                  employeeNickname: "닉네임"
                  contractStartTime: "2025-01-01T09:00:00"
                  contractEndTime: "2025-01-01T18:00:00"
                  reviewStarPoint: 5
                  reviewContent: "훌륭한 작업이었습니다!"
                - reviewId: 2
                  businessName: "스타트업 컴퍼니"
                  businessId: 20
                  employeeId: 3
                  employeeNickname: "닉네임"
                  contractStartTime: "2025-01-05T10:00:00"
                  contractEndTime: "2025-01-05T19:00:00"
                  reviewStarPoint: 4
                  reviewContent: "업무 진행이 원활했습니다."
  /api/v1/contracts/review/modify:
    put:
      tags:
        - ContractReviewController
      summary: "리뷰 수정하기"
      description: "리뷰 수정하기"
      operationId: modifyReview
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModifyReviewRequest"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              example: "성공적으로 삭제되었습니다."
  /api/v1/contracts/review/delete:
    delete:
      tags:
        - ContractReviewController
      summary: "리뷰 삭제하기"
      description: "리뷰 삭제하기"
      operationId: modifyReview
      parameters:
        - name: reviewId
          in: query
          required: true
          schema:
            type: integer
            format: int64
            example: 1
      responses:
        "200":
          description: OK
          content:
            '*/*':
              example: "성공적으로 삭제되었습니다."
...
  /api/v1/review:
    get:
      tags:
        - ReviewController
      operationId: getAllReviews
      summary: "알바 후기 전체 조회"
      parameters:
        - name: nickname
          in: query
          required: false
          schema:
            type: string
            example: "단팥"
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int64
            example: 1
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReviewResponse"
                example:
                  - reviewId: 1
                    businessName: "테스트 사업장"
                    businessId: 1
                    employeeId: 2
                    employeeNickname: "닉네임1"
                    contractStartTime: "2025-01-01T09:00:00"
                    contractEndTime: "2025-01-01T18:00:00"
                    reviewStarPoint: 5
                    reviewContent: "훌륭한 작업이었습니다!"
                  - reviewId: 2
                    businessName: "또 다른 사업장"
                    businessId: 20
                    employeeId: 3,
                    employeeNickname: "닉네임2"
                    contractStartTime: "2025-01-02T09:00:00"
                    contractEndTime: "2025-01-02T18:00:00"
                    reviewStarPoint: 4
                    reviewContent: "좋은 성과를 보여주었습니다."
  /api/v1/review/{reviewId}:
    get:
      tags:
        - ReviewController
      operationId: getReviewsByEmployee
      summary: "특정 알바 후기 조회"
      parameters:
        - name: reviewId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReviewResponse"
                example:
                  - reviewId: 1
                    businessName: "테스트 사업장"
                    businessId: 10
                    contractStartTime: "2025-01-01T09:00:00"
                    contractEndTime: "2025-01-01T18:00:00"
                    reviewStarPoint: 5
                    reviewContent: "훌륭한 작업이었습니다!"
  /api/v1/review/my/employee:
    get:
      tags:
        - ReviewController
      summary: "나를(알바생) 대상으로 작성된 리뷰 보기"
      operationId: getMyReviews
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int64
            example: 1
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReviewResponse"
                example:
                  - reviewId: 1
                    businessName: "테스트 사업장"
                    businessId: 10
                    employeeId: 3
                    employeeNickname: "닉네임"
                    contractStartTime: "2025-01-01T09:00:00"
                    contractEndTime: "2025-01-01T18:00:00"
                    reviewStarPoint: 5
                    reviewContent: "훌륭한 작업이었습니다!"
                  - reviewId: 2
                    businessName: "스타트업 컴퍼니"
                    businessId: 20
                    employeeId: 3
                    employeeNickname: "닉네임"
                    contractStartTime: "2025-01-05T10:00:00"
                    contractEndTime: "2025-01-05T19:00:00"
                    reviewStarPoint: 4
                    reviewContent: "업무 진행이 원활했습니다."
  /api/v1/review/available:
    get:
      tags:
        - ReviewController
      summary: "작성할 수 있는 리뷰 목록 보기"
      operationId: getAvailableReviewTargets
      parameters:
        - name: businessId
          in: query
          required: true
          schema:
            type: integer
            format: int64
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int64
            example: 1
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReviewAvailableResponse"
                example:
                  - employeeId: 1001
                    employeeName: "김철수"
                  - employeeId: 1002
                    employeeName: "이영희"
...
components:
  schemas:
...
    AddReviewRequest:
      required:
      - contractId
      type: object
      properties:
        contractId:
          type: integer
          format: int64
        businessId:
          type: integer
          format: int64
        employeeId:
          type: integer
          format: int64
        reviewScore:
          type: integer
          format: int32
        reviewContent:
          type: string
...
    ReviewResponse:
      type: object
      properties:
        reviewId:
          type: integer
          format: int64
        businessName:
          type: string
        businessId:
          type: integer
          format: int64
        contractStartTime:
          type: string
          format: date-time
        contractEndTime:
          type: string
          format: date-time
        reviewStarPoint:
          type: integer
          format: int32
        reviewContent:
          type: string
    ReviewAvailableResponse:
      type: object
      properties:
        employeeId:
          type: integer
          format: int64
        employeeName:
          type: string
...
    ModifyReviewRequest:
      type: object
      properties:
        reviewId:
          type: integer
          example: 2
        reviewScore:
          type: integer
          example: 5
        reviewContent:
          type: string
          example: "일 잘하네여"
  securitySchemes:
    JWT_TOKEN:
      type: http
      scheme: bearer
      bearerFormat: JWT